+++ sandbox/k8s/tf-sourceloop-sandbox/example/main.tf
@@ -1,0 +1,5 @@
+module "example_module" {
+  source = "../."
+
+  mandatory = "mandatory_string"
+}